requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
ci-info==0.3.0
clarifai-grpc==9.7.4
click==8.1.7
configobj==5.0.8
configparser==6.0.0
etelemetry==0.3.0
filelock==3.12.2
fitz==0.0.1.dev2
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
googleapis-common-protos==1.60.0
grpcio==1.57.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
looseversion==1.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathlib==1.0.1
Pillow==9.5.0
protobuf==4.24.2
prov==2.0.0
pyarrow==13.0.0
pydeck==0.8.0
pydot==1.4.2
Pygments==2.16.1
Pympler==1.0.1
PyMuPDF==1.23.1
PyMuPDFb==1.23.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.6
rdflib==7.0.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scipy==1.11.2
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traits==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
zipp==3.16.2
